@@ -1,9 +1,9 @@
 class UsersController < Devise::OmniauthCallbacksController
   include Users::PassiveLogin
-  prepend_before_filter :redirect_root, only: [:show], if: -> { request.path == '/' && user_signed_in? }
-  before_filter :require_login, only: [:show]
-  before_filter :require_no_authentication, except: [:passthru, :show, :client_passive_login]
-  before_filter :require_valid_omniauth_hash, only: (Devise.omniauth_providers << :omniauth_callback)
+  prepend_before_action :redirect_root, only: [:show], if: -> { request.path == '/' && user_signed_in? }
+  before_action :require_login, only: [:show]
+  before_action :require_no_authentication, except: [:passthru, :show, :client_passive_login]
+  before_action :require_valid_omniauth_hash, only: (Devise.omniauth_providers << :omniauth_callback)
   respond_to :html
 
   def show
