@@ -1,7 +1,7 @@
 class WayfController < ApplicationController
   respond_to :html
   # Don't allow users to visit the logged_out page with GET, force a logout if that happens
-  before_filter :redirect_to_logout, only: [:logged_out], if: -> { user_signed_in? }
+  before_action :redirect_to_logout, only: [:logged_out], if: -> { user_signed_in? }
 
   def index
     if cookies[:institution_from_url].present?
