@@ -1,13 +1,10 @@
-class Identity < ActiveRecord::Base
+class Identity < ApplicationRecord
   VALID_PROVIDERS = Devise.omniauth_providers.map(&:to_s)
 
   belongs_to :user
 
   # Must have a uid, provider and properties
-  validates :uid, presence: true
-  validates :provider, presence: true
-  # Properties is an Hstore column type
-  validates :properties, presence: true
+  validates :uid, :provider, :properties, presence: true
 
   # Must have a unique uid per provider/user
   validates :uid, uniqueness: { scope: [:user_id, :provider] }
